requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
pytz>=2023.3